fix: removed radios from admin page

+++ services/AdminService.js
@@ -39,8 +39,10 @@ class AdminService {
 
     async adminDeletePost(id) {
         try {
-            return await AdminRepository.deleteAdmin(id);
-
+            var admins = await AdminRepository.findAll();
+            if (Object.keys(admins).length > 1) {
+                await AdminRepository.deleteAdmin(id);
+            }
         } catch (error) {
             return error;
         }
@@ -57,58 +59,53 @@ class AdminService {
 
     }
 
-    async adminUpdatePost(id, login, password, radio) {
+    async adminUpdatePost(id, login, password) {
         try {
             var collaborator = await AdminRepository.findCollaboratorByLogin(login);
-            if (collaborator == null) {
-                var admin = await AdminRepository.findByPk(id);
-                var login_name = await AdminRepository.findOne(login);
-                if (radio == "option1") {
-                    if (login_name == undefined) {
-                        await AdminRepository.updateAdminLogin(login, id);
+            var login_name = await AdminRepository.findOne(login, id);
 
-                        admin = {
-                            id: id,
-                            login: login
-                        }
-                        return 1;
-                    } else {
-                        return 0;
+            if (collaborator == null && login_name == null) {
+                var admin = await AdminRepository.findByPk(id);
+
+                if (password == "") {
+                    await AdminRepository.updateAdminLogin(login, id);
+                    admin = {
+                        id: id,
+                        login: login
                     }
+                    return 1;
 
                 } else {
-                    if (password != "") {
-                        var salt = bcrypt.genSaltSync(10);
-                        var hash = bcrypt.hashSync(password, salt);
-                        await AdminRepository.adminUpdatePassword(hash, id);
-                        return admin;
-                    } else {
-                        return 0;
+                    var salt = bcrypt.genSaltSync(10);
+                    var hash = bcrypt.hashSync(password, salt);
+                    await AdminRepository.adminUpdatePassword(hash, id);
+                    await AdminRepository.updateAdminLogin(login, id);
+                    admin = {
+                        id: id,
+                        login: login
                     }
+                    return 1;
                 }
-            } else {
-                return -1;
             }
-
         } catch (error) {
             return error;
         }
     }
-
-    /* 
-        async createAdminInitialProfile() {
-            var admins = await AdminRepository.findAll();
-            if (admins == null) {
-    
-                var login = 'admin';
-                var password = 'admin';
-    
-                var salt = bcrypt.genSaltSync(10);
-                var hash = bcrypt.hashSync(password, salt);
-                await AdminRepository.createAdmin(login, hash);
-            }
-        } */
-
 }
+
+/* 
+    async createAdminInitialProfile() {
+        var admins = await AdminRepository.findAll();
+        if (admins == null) {
+ 
+            var login = 'admin';
+            var password = 'admin';
+ 
+            var salt = bcrypt.genSaltSync(10);
+            var hash = bcrypt.hashSync(password, salt);
+            await AdminRepository.createAdmin(login, hash);
+        }
+    } */
+
 
 module.exports = new AdminService();
